Fix expansion of module with module type path to module types
Eg in:

```
module X : A.B.Y
```

The module type of `X` is the path to module type `A.B.Y`.

This is represented in `Types` as `Mty_ident`. Before this commit it was treated
as `Mty_alias` which represents module aliases (`module X = A.B.Y` where `A.B.Y`
has to be a path to a module)